% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Authors_stagflation.R
\docType{data}
\name{Authors_stagflation}
\alias{Authors_stagflation}
\title{List Of Authors Of The Articles and Books Explaining the 1970s US Stagflation.}
\format{
A data frame with 231 rows and 3 variables:
\describe{
\item{source_id}{Identifier of the document published by the author}
\item{author_name}{Author of the document}
\item{author_order}{Author order in the document author list}
}
}
\source{
Goutsmedt A. (2020) “From Stagflation to the Great Inflation: Explaining the 1970s US Economic
Situation”. Revue d’Economie Politique, Forthcoming 2021.
}
\usage{
Authors_stagflation
}
\description{
A dataset associating the books and academic articles endeavouring to explain
the US stagflation and their authors (\code{Nodes_stagflation} just takes the first author;
here is the complete list of authors per document).
}
\keyword{datasets}
