% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_network.R
\name{color_network}
\alias{color_network}
\title{Color the nodes and edges of the network according to a given variable.}
\usage{
color_network(graphs = NA, column_to_color = NA, color = NA)
}
\arguments{
\item{column_to_color}{This column designates the categorical variable you want to use to color your nodes.}

\item{color}{The color you want to use in your networks.
It can be a vector of colors or a two columns data.frame matching the first column as the distinct observations of the column_to_color and a second column with the vector of colors you want to use}

\item{graph}{A network or a list of networks.}
}
\description{
This function will create a color column for the edges and nodes in your network. You can chose a column with a categorical variable
to color nodes based on this particular variable. Edges will be colored by mixing the color of connected nodes.
By default, the function will color your network by using 19 colors from the RColorBrewer package, but you can chose your own color.
}
