% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/community_colors.R
\name{community_colors}
\alias{community_colors}
\title{Creating Color Attribute For Network Communities}
\usage{
community_colors(graph, palette, community_column = "Com_ID")
}
\arguments{
\item{graph}{A tidygraph graph}

\item{palette}{The palette to be used for attributing colors to communities. If you use a palette with less color
than the total number of communities, you will have communities represented by the same color in your visualisation.}

\item{community_column}{The name of the column with the community identifier/number. By defautl "Com_ID", as it is
the name of the column when you use the \code{\link[=leiden_workflow]{leiden_workflow()}} function for detecting communities.}
}
\value{
The same tidygraph object but with a new \code{color} column in the nodes side, for each community, and a new
\code{color_edges} column in the edges side.
}
\description{
This function takes as an input a tidygraph graph, with a column with a community identifier for
each node and each edge. It attributes to each community a color, depending on a palette chosen.
If the two nodes connected by an edge have a different community, the function mixes the color of the two communities.
}
