% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/community_labels.R
\name{community_labels}
\alias{community_labels}
\title{Calculating the Coordinates for Projecting Community Labels}
\usage{
community_labels(
  graph,
  biggest_community = FALSE,
  community_threshold = 0.01,
  community_name_column = "Community_name",
  community_size_column = "Size_com"
)
}
\arguments{
\item{graph}{A tidygraph object.}

\item{biggest_community}{If true, you have the possibility to remove the smallest community, depending of the \code{community_threshold}
you have set.}

\item{community_threshold}{If \code{biggest_community} is true, the function selects the nodes that belong to communities which represent
at least x\% of the total number of nodes. By default, the parameter is set to 1\%.}

\item{community_name_column}{Name of the column with the name of the community to be used as label}

\item{community_size_column}{Name of the column with the total number of nodes or the share of nodes in each community.}
}
\value{
A data.frame with the list of your communities names (only the biggest one if \code{biggest_community}
set to \code{TRUE}) and a pair of coordinates to project these names on your visualisation. It also
keeps the color of the community and its size.
}
\description{
A simple function to calculate the mean of coordinates x and y for each community. These coordinates
are to be used to plot the name of the community on the graph visualisation.
}
\details{
You have to run this function only after having calculating (x,y) coordinates for your nodes.
}
