% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/community_names.R
\name{community_names}
\alias{community_names}
\title{Automatically Attributing Names to Communities}
\usage{
community_names(
  graph,
  ordering_column,
  naming = "Label",
  community_column = "Com_ID"
)
}
\arguments{
\item{graph}{A tidygraph object.}

\item{ordering_column}{Enter the name of the column you want to be used to choose the community name. For instance, if
you choose \code{Degree}, the function takes the value of the \code{naming} column of the node with the highest
degree in the community to name the community. You can use other measure
than network centrality measures: for instance, if nodes are articles, you can use the number of citations of articles.}

\item{naming}{Enter the name of the column you want to be used for naming the community. The function takes the node
with the highest \code{centrality_measure} chosen, and use the node value in the \code{naming} column  to title the community.
For instance, if nodes are individuals and if you have a column called \code{surname}, you can use this column.}

\item{community_column}{The name of your community identifiers column.}
}
\value{
The same graph object but with a column \code{Community_name}, as well as \code{Community_2_name}
and \code{Community_3_name} if you have run the \code{\link[=leiden_workflow]{leiden_workflow()}} function for more than one resolution.
}
\description{
A function to give to a community the name of its node with the highest chosen measure.
It also gives the edges the name of their community. If the edge connects nodes from
different community, name will be NA.

The function takes into account the parameters chosen for the \code{leiden_improved()} function.
If you have chosen 2 or 3 levels of resolution, it repeats the same process for the second and third resolution.
In other words, for 3 levels of resolution, you will have the names of the communities for the first value of the
Leiden resolution, but also the names for the second and third values.
}
\details{
The attribute of nodes and edges with the names of the communities is called
\code{Community_name}. It is formed by the identifier of the community (in the \code{Com_ID} column)
and by the \code{naming} value.

If you have entered a second and a third resolutions values in the
\code{leiden_improved()} function, you will have two supplementary columns: \code{Community_2_name}
and \code{Community_3_name}.
}
