% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_components.R
\name{filter_components}
\alias{filter_components}
\title{Filtering Network Components}
\usage{
filter_components(
  graphs,
  nb_components = 1,
  threshold_alert = 0.05,
  keep_component_columns = FALSE
)
}
\arguments{
\item{graphs}{A tidygraph object or a list of tidygraph objects with a "weight" column
for edges.}

\item{nb_components}{By default, the function takes the main component of the graph (\code{nb_components = 1}).
However it is possible to take as many components as you wish. The first component
is the largest one, component 2 is the second one, etc.}

\item{threshold_alert}{If the biggest component after the last one selected (by default, nb_component = 1)
gathers more than x\% (by default, 5\%) of the total number of nodes,
the function triggers a warning to inform the user that he has removed a big component of the network.}

\item{keep_component_columns}{Set to \code{TRUE} if you want to store in the tibble graph the components number and
the size of the components.}
}
\value{
The same tidygraph object or list of tidygraph objects with nodes
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This function is deprecated and will be replaced by \code{networkflow::extract_main_component()}.
This function which i) creates a \href{https://tidygraph.data-imaginist.com/index.html}{tidygraph}
graph using \link[tidygraph:tbl_graph]{tbl_graph()};
ii) keeps the main components of the graph, using \link[tidygraph:group_graph]{main_components()}; and iii) warns
the user if the first biggest component removed is too large.
}
\details{
The function will automatically rename the first column of nodes as "Id".
}
