% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_network_app.R
\name{launch_network_app}
\alias{launch_network_app}
\title{Launch an Interactive Shiny App to Explore a Network Graph}
\usage{
launch_network_app(
  graph_tbl,
  cluster_id,
  cluster_information,
  cluster_tooltip = NULL,
  node_id,
  node_tooltip = NULL,
  node_size = NULL,
  color = NULL,
  layout = "kk"
)
}
\arguments{
\item{graph_tbl}{A \code{tbl_graph} object with or without layout coordinates.}

\item{cluster_id}{Column name in the node data identifying clusters. It can be the cluster ID or a unique label for each cluster.}

\item{cluster_information}{Character vector of node metadata columns to display in the data table.}

\item{cluster_tooltip}{Optional. Text string to display when hovering over cluster labels.}

\item{node_id}{Column name identifying node IDs.}

\item{node_tooltip}{Optional. Column name for tooltips shown when hovering over nodes.}

\item{node_size}{Optional. Column name used to scale node size.}

\item{color}{Optional. Column name used to color nodes. If \code{NULL}, colors are automatically assigned using \code{networkflow::color_networks(column_to_color = cluster_id)}.}

\item{layout}{Character. Name of a layout to compute node coordinates (e.g. \code{"kk"}, \code{"fr"}). If specified, this replaces any existing layout.}
}
\value{
A Shiny app interface for exploring the network.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This function launches an interactive Shiny application for visualizing and exploring a network graph represented as a \code{tbl_graph} object.
It supports node and cluster interactivity with tooltips, dynamic sizing, and customizable coloring.
Users can inspect clusters interactively and explore associated node metadata in a searchable datatable.

If the graph does not contain a layout (columns \code{x} and \code{y}) the function will compute coordinates using
\code{networkflow::layout_networks()} with the specified layout algorithm.  If a layout name is provided, it will replace any existing layout in the graph.
}
\examples{
library(networkflow)
library(dplyr)

nodes <- Nodes_stagflation |>
  dplyr::filter(Type == "Stagflation") |>
  dplyr::rename(ID_Art = ItemID_Ref)

references <- Ref_stagflation |>
  dplyr::rename(ID_Art = Citing_ItemID_Ref)

g <- build_network(
  nodes = nodes,
  directed_edges = references,
  source_id = "ID_Art",
  target_id = "ItemID_Ref",
  cooccurrence_method = "coupling_similarity",
  edges_threshold = 1,
  compute_size = FALSE,
  keep_singleton = FALSE
)

g <- add_clusters(
  g,
  clustering_method = "leiden",
  objective_function = "modularity"
)

launch_network_app(
  graph_tbl = g,
  cluster_id = "cluster_leiden",
  cluster_information = c("Author", "Title", "Year", "Journal"),
  cluster_tooltip = "Cluster",
  node_id = "ID_Art",
  node_tooltip = "Author_date",
  node_size = NULL,
  color = NULL,
  layout = "kk"
)

# Dynamic networks
g_list <- build_dynamic_networks(
  nodes = nodes,
  directed_edges = references,
  source_id = "ID_Art",
  target_id = "ItemID_Ref",
  time_variable = "Year",
  time_window = 20,
  cooccurrence_method = "coupling_similarity",
  edges_threshold = 1,
  overlapping_window = TRUE,
  compute_size = FALSE,
  keep_singleton = FALSE
)

g_list <- add_clusters(
  g_list,
  clustering_method = "leiden",
  objective_function = "modularity"
)

launch_network_app(
  graph_tbl = g_list,
  cluster_id = "cluster_leiden",
  cluster_information = c("Author", "Title", "Year", "Journal"),
  node_id = "ID_Art",
  node_tooltip = "Author_date",
  node_size = NULL,
  color = NULL,
  layout = "kk"
)
}
