% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimize_crossing_alluvial.R
\name{minimize_crossing_alluvial}
\alias{minimize_crossing_alluvial}
\title{Order the rectangles of the alluvial in a way that minimize crossing of flow}
\usage{
minimize_crossing_alluvial(
  alluv_dt = NA,
  cluster_column = NA,
  node_key = NA,
  time_variable = NA
)
}
\arguments{
\item{alluv_dt}{Data.frame of the alluvial created using the networkflow::networks_to_alluv function}

\item{cluster_column}{The column with the identifier of the clusters. This is stratum of the alluvial.}

\item{node_key}{The column with the unique identifier of each node. This is the alluvium of the alluvial.}

\item{time_variable}{The column with the temporal variable you used to build your windows or your list of networks.

#' @examples}
}
\description{
This function transform the node_key into a factor ordered by a newly created column named "minimize_crossing_order".
}
