% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkflow-package.R
\docType{package}
\name{networkflow-package}
\alias{networkflow}
\alias{networkflow-package}
\title{networkflow: Functions For A Workflow To Manipulate Networks}
\description{
Provides a workflow to build, analyze, and visualize projected networks from tabular data. The package supports dynamic analysis across time windows, including cluster detection and cross-period cluster matching. It also covers network construction, interpretation, static plotting, and interactive exploration through a 'shiny' app. Although designed for projected networks (e.g., article -> reference), it can be used more generally with 'tbl_graph' objects.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/agoutsmedt/networkflow}
  \item \url{https://agoutsmedt.github.io/networkflow/}
  \item Report bugs at \url{https://github.com/agoutsmedt/networkflow/issues}
}

}
\author{
\strong{Maintainer}: Aurélien Goutsmedt \email{agoutsmedt@gmail.com} (\href{https://orcid.org/0000-0002-3788-7237}{ORCID})

Authors:
\itemize{
  \item Alexandre Truc \email{alexandre.truc77@gmail.com} (\href{https://orcid.org/0000-0002-1328-7819}{ORCID})
  \item Thomas Delcey (\href{https://orcid.org/0000-0003-0546-1474}{ORCID})
}

}
\keyword{internal}
