% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nodes_stagflation.R
\docType{data}
\name{Nodes_stagflation}
\alias{Nodes_stagflation}
\title{Articles and Books Explaining the 1970s US Stagflation.}
\format{
A data frame with 654 rows and 7 variables:
\describe{
\item{source_id}{Identifier of the document}
\item{source_author}{Author of the document}
\item{source_label}{Use this as a label for nodes}
\item{source_year}{Year of publication of the document}
\item{source_title}{Title of the document}
\item{source_journal}{Journal of publication of the document (if an article)}
\item{source_type}{If "Stagflation", the document is listed as an explanation of the US stagflation.
If "Non-Stagflation", the document is cited by a document explaining the stagflation}
}
}
\source{
Goutsmedt A. (2020) “From Stagflation to the Great Inflation: Explaining the 1970s US Economic
Situation”. Revue d’Economie Politique, Forthcoming 2021.
}
\usage{
Nodes_stagflation
}
\description{
A dataset containing the books and academic articles endeavouring to explain
what happened in the US economy in the 1970s, as well as all the articles and books
cited at least twice by the first set of articles and books (on the stagflation).
}
\keyword{datasets}
