% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_networks.R
\name{plot_networks}
\alias{plot_networks}
\title{Plot your network or list of networks after having followed the workflow of networkflow.}
\usage{
plot_networks(
  graphs = NA,
  networkflow_label_column = TRUE,
  label_column = NA,
  x_coordinates_column = "x",
  y_coordinates_column = "y",
  color_nodes_column = "color",
  color_edges_column = "color_edges",
  weight_column = "weight",
  node_size_column = "size",
  time_variable = NA
)
}
\arguments{
\item{graphs}{A network or a list of networks.}

\item{networkflow_label_column}{Do you want to use the label created with the \link[=label_networks]{label_networks()} function ? TRUE or FALSE.}

\item{label_column}{If networkflow_label_column is TRUE, this parameters is not used. If you want to plot labels for individual nodes you can use this parameters to chose the column to use as label.}

\item{x_coordinates_column}{The column with the x coordinates of your nodes.}

\item{y_coordinates_column}{The column with the y coordinates of your nodes.}

\item{color_nodes_column}{The column with the colors of your nodes.}

\item{color_edges_column}{The column with the colors of your edges.}

\item{weight_column}{The column with the weight of edges.}

\item{node_size_column}{The column with the size of nodes.}

\item{time_variable}{The column with the temporal variable of your individual nodes. Will only be used for the title of the plot.}
}
\description{
An easy way to plot the your networks by using the columns created by the different function of networkflow.
}
