% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_label_networks.R
\name{prepare_label_networks}
\alias{prepare_label_networks}
\alias{prepare_label_alluvial}
\title{Creating and Positioning Labels for Plotting Networks or Alluvial}
\usage{
prepare_label_networks(
  graphs,
  x = "x",
  y = "y",
  cluster_label_column = "cluster_label"
)

prepare_label_alluvial(
  alluv_dt = NA,
  window_column = "window",
  cluster_label_column = "cluster_label"
)
}
\arguments{
\item{graphs}{A tibble graph from \href{https://tidygraph.data-imaginist.com/}{tidygraph} or a list of tibble
graphs.}

\item{x, y}{The columns of your tibble graph or list of tibble graphs
with the x and y coordinates of your nodes (for instance, the coordinates
produce by \link[=layout_networks]{layout_networks()}).}

\item{cluster_label_column}{The column with the cluster label you want to display. By default, "cluster_label", as it is
the default name of the column created with \link[=name_clusters]{name_clusters()}.
You may also use the cluster identifier created by \link[=add_clusters]{add_clusters()}
for simple cluster, or by \link[=merge_dynamic_clusters]{merge_dynamic_clusters()} for
inter-temporal clusters.}

\item{alluv_dt}{A data.frame of an alluvial created with \link[=networks_to_alluv]{networks_to_alluv()}.}

\item{window_column}{The column of the alluvial with your time window. By default, "window", as created by
\link[=networks_to_alluv]{networks_to_alluv()}.}
}
\description{
\code{prepare_label_networks()} creates a label column that make it easy to center label around particular clusters on the alluvial graph of intertemporal networks.
\code{prepare_label_alluvial()}
}
