% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ref_stagflation.R
\docType{data}
\name{Ref_stagflation}
\alias{Ref_stagflation}
\title{Articles and Books Explaining the 1970s US Stagflation.}
\format{
A data frame with 4416 rows and 6 variables:
\describe{
\item{source_id}{Identifier of the citing document}
\item{target_id}{Identifier of the cited document}
\item{target_author}{Author of the cited document}
\item{target_year}{Year of publication of the cited document}
\item{target_title}{Title of the cited document}
\item{target_journal}{Journal of publication of the cited document (if an article)}
}
}
\source{
Goutsmedt A. (2020) “From Stagflation to the Great Inflation: Explaining the 1970s US Economic
Situation”. Revue d’Economie Politique, Forthcoming 2021.
}
\usage{
Ref_stagflation
}
\description{
A dataset containing all the articles and books cited by
the books and academic articles endeavouring to explain
what happened in the US economy in the 1970s.
}
\keyword{datasets}
