% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_main_component.R
\name{tbl_main_component}
\alias{tbl_main_component}
\title{Creating Graph And Keeping Main Component With Tidygraph From Edges And Nodes}
\usage{
tbl_main_component(
  edges,
  nodes,
  nb_components = 1,
  threshold_alert = 0.05,
  ...
)
}
\arguments{
\item{edges}{A dataframe with a list of links between nodes, under the columns "from" and "to". The two columns should
be in character format.}

\item{nodes}{A dataframe with a list of nodes. The first column will be used as the identifying column.
Be careful to avoid doublons in the first column. The first column should be in character format.}

\item{nb_components}{By default, the function takes the main component of the graph (nb_components = 1). However it is possible to take as
many components as you wish Component 1 is the largest one, component 2 is the second one, etc.}

\item{threshold_alert}{If the biggest component after the last one selection (by default, nb_component = 1) gathers more than x\% (by default, 5\%) of the total number of nodes,
the function triggers a warning to inform the user that he has removed a big component of the network.}

\item{...}{Any parameter of the \link[tidygraph:tbl_graph]{tbl_graph()} function. Edges are directed by default.
Coupling or co-citation network are \emph{undirected} networks.}
}
\value{
A tidygraph object with a list of nodes and a list of edges.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function is deprecated and will be replaced by \code{networkflow::extract_main_component()}.
This function which i) creates a \href{https://tidygraph.data-imaginist.com/index.html}{tidygraph}
graph using \link[tidygraph:tbl_graph]{tbl_graph()};
ii) keeps the main components of the graph, using \link[tidygraph:group_graph]{main_components()}; and iii) warns
the user if the first biggest component removed is too large.
}
\details{
The function will automatically rename the first column of nodes as "Id".
}
