% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_tfidf.R
\name{tf_idf}
\alias{tf_idf}
\title{Extracting TF-IDF Values for Ngrams}
\usage{
tf_idf(
  data,
  text_columns,
  grouping_columns,
  n_gram = 2L,
  stopwords = NULL,
  stopwords_list = "smart",
  clean_word_method = c("lemmatize", "stemming", "none"),
  ngrams_filter = 5L
)
}
\arguments{
\item{data}{A \href{https://tidygraph.data-imaginist.com/}{tidygraph} object, a list of tidygraph objects or a simple table.}

\item{text_columns}{The columns with the text you want to analyze. If you give multiple columns, they will
be gathered to extract the terms.}

\item{grouping_columns}{The columns you want to use to calculate the tf-idf. These columns will become your
"document" unit in the \link[tidytext:bind_tf_idf]{bind_tf_idf()} function.}

\item{n_gram}{The maximum n you want for tokenizing your ngrams (see \link[tidytext:unnest_tokens]{unnest_tokens()}
for more information). 2 by default, i.e. only unigrams and bigrams will be extracted.}

\item{stopwords}{Use your own stopwords list, in a character vector format}

\item{stopwords_list}{The stopwords list you want to use to remove stopwords from your ngrams. The "smart" list by
default, but see other possilities with \link[stopwords:stopwords_getsources]{stopwords_getsources}.}

\item{clean_word_method}{Choose the method to clean and standardized your ngrams. You can lemmatize or stem words through
the \href{https://github.com/trinker/textstem}{textstem package}. Choose "none" if you don't want to apply any cleaning method.}

\item{ngrams_filter}{You can exclude from tf-idf computation the ngrams that does not appear a certain number of time.}
}
\value{
A data.table with the list of terms (i.e. ngrams) appearing in each "document" (that is your
\code{grouping_columns}) with the number of time they appear per document (\code{n}), their
term frequency (\code{tf}), their inverse document frequency (\code{idf}), and their term-frequency inverse-document-frequency
(\code{tf_idf}).
}
\description{
This function takes as input a tidygraph object, a list of tidygraph objects or a data frame,
extract the ngrams use in the character columns of your choice, and calculates the Term-Frequency
Inverse-Document-Frequency value of each ngram for each document.
}
