% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Edges_coupling.R
\docType{data}
\name{Edges_coupling}
\alias{Edges_coupling}
\title{Edges For Bibliographic Coupling Network Of Articles and Books Explaining the 1970s US Stagflation.}
\format{
A data frame with 154 rows and 6 variables:
\describe{
\item{from}{Identifier of the Source document on stagflation, in character format}
\item{to}{Identifier of the Target document on stagflation, in character format}
\item{weight}{weight calculated according to the "coupling_angle" method (see \link[biblionetwork:biblio_coupling]{biblio_coupling()})}
\item{Source}{copy of the identifiers of the Source document on stagflation}
\item{Target}{copy of the identifiers of the Target document on stagflation}
}
}
\source{
Created from \code{Nodes_stagflation.rda}
}
\usage{
Edges_coupling
}
\description{
A dataset containing the edges of the bibliographic coupling network of articles and books on stagflation.
Built by using \link{Ref_stagflation}: \code{biblionetwork::biblio_coupling(Ref_stagflation,"Citing_ItemID_Ref","ItemID_Ref")}.
Could be used with \link{Nodes_coupling} to create a network with tidygraph.
}
\keyword{datasets}
