% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout_networks.R
\name{layout_networks}
\alias{layout_networks}
\title{Calculate the Coordinates of Nodes}
\usage{
layout_networks(
  graphs,
  node_id,
  layout,
  compute_dynamic_coordinates = FALSE,
  save_coordinates = FALSE,
  ...
)
}
\arguments{
\item{graphs}{A tibble graph from \href{https://tidygraph.data-imaginist.com/}{tidygraph}, a list of tibble
graphs or a data frame.}

\item{node_id}{The column with the unique identifier of each node.}

\item{layout}{The type of layout to create. All strings accepted by the algorithm argument can
also be supplied directly into layout. See \code{ggraph::ggraph()} for more details on
layouts.}

\item{compute_dynamic_coordinates}{When you have a list of tibble graphs and that some nodes are in multiple graphs,
you may want that each node in several graphs has relatively similar coordinates. Set
\code{compute_dynamic_coordinates} to \code{TRUE} to take into account the coordinates of the nodes in
the \code{n-1} tibble graph, during the computation of the coordinates of the \code{n} tibble graph.}

\item{save_coordinates}{If you are running the function on the same object, the already existing columns \code{x} and \code{y}
will be erased and new values will be computed. Set \code{save_coordinates} to \code{TRUE} to
save the \code{x} and \code{y} values in columns with a name depending of the layout method used.}

\item{...}{Additional arguments of the layout compatible with the function. See \code{ggraph::graph()},
\code{ggraph::layout_tbl_graph_igraph()} and \code{graphlayouts::graphslayouts()} for more information.}
}
\value{
The same tibble graph or list of tibble graphs with a column \code{x} and \code{y}, and also
additional column \verb{x_\{layout\}} and \verb{y_\{layout\}} if you have set
\code{save_coordinates} to \code{TRUE}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

The function implements different layout algorithms from
\href{https://ggraph.data-imaginist.com/index.html}{ggraph} and
\href{http://graphlayouts.schochastics.net/index.html}{graphlayouts}. It also implements
a dynamic process of computation of coordinates for list of tibble graphs, in order
for the coordinates to be consistent between graphs.
}
\details{
The function allows for the use of different layouts implemented in
\href{https://ggraph.data-imaginist.com/index.html}{ggraph} and
\href{http://graphlayouts.schochastics.net/index.html}{graphlayouts}. However, some
layout of \code{graphlayouts} are not working with this function. Also, layouts do not
all take as input pre-determined coordinates, meaning that the
\code{compute_dynamic_coordinates} argument does not work with all layout. Please check
if the layout used allows a parameter called \code{coord}.
}
\examples{
library(networkflow)

nodes <- Nodes_stagflation |>
dplyr::rename(ID_Art = ItemID_Ref) |>
dplyr::filter(Type == "Stagflation")

references <- Ref_stagflation |>
dplyr::rename(ID_Art = Citing_ItemID_Ref)

temporal_networks <- build_dynamic_networks(nodes = nodes,
directed_edges = references,
source_id = "ID_Art",
target_id = "ItemID_Ref",
time_variable = "Year",
cooccurrence_method = "coupling_similarity",
time_window = 20,
edges_threshold = 1,
overlapping_window = TRUE,
filter_components = TRUE)

temporal_networks <- layout_networks(temporal_networks,
node_id = "ID_Art",
layout = "fr",
compute_dynamic_coordinates = TRUE)

temporal_networks [[1]]

}
