% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimize_crossing_alluvial.R
\name{minimize_crossing_alluvial}
\alias{minimize_crossing_alluvial}
\title{Order the rectangles of the alluvial in a way that minimize crossing of flow}
\usage{
minimize_crossing_alluvial(
  alluv_dt = NULL,
  intertemporal_cluster_column = NULL,
  node_key = NULL,
  window_column = "Window"
)
}
\arguments{
\item{alluv_dt}{Data.frame of the alluvial created using the networkflow::networks_to_alluv function}

\item{intertemporal_cluster_column}{The column with the identifier of the inter-temporal cluster. If you have used
\link[=add_clusters]{add_clusters()} and \link[=merge_dynamic_clusters]{merge_dynamic_clusters()},
it is of the form \verb{dynamic_cluster_\{clustering_method\}}.}

\item{node_key}{The column with the unique identifier of each node. This is the alluvium of the alluvial.}

\item{window_column}{The column with your time windows.

@examples}
}
\description{
This function transform the node_key into a factor ordered by a newly created column named "minimize_crossing_order".
}
