% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_clusters.R
\name{name_clusters}
\alias{name_clusters}
\title{Automatically Attributing Names to Clusters}
\usage{
name_clusters(
  graphs,
  method = c("tidygraph_functions", "given_column", "tf-idf"),
  name_merged_clusters = FALSE,
  cluster_id,
  label_columns,
  label_name = "cluster_label",
  tidygraph_function = NULL,
  order_by = NULL,
  text_columns = NULL,
  nb_terms_label = 3,
  ...
)
}
\arguments{
\item{graphs}{A tibble graph (from \href{https://tidygraph.data-imaginist.com/}{tidygraph})
or a list of tibble graphs.}

\item{method}{The method for finding the names, among \code{tidygraph_functions}, \code{given_column},
and \code{tf-idf} (see the details). The \code{tf-idf} method is chosen by default.}

\item{name_merged_clusters}{Set to \code{TRUE} if your clusters have been established for
all your tibble graphs and thus are unique. Typically, you have such clusters
after running \code{merge_dynamic_clusters()}.}

\item{cluster_id}{The column you want to name. Generally, the column with the identifier of the
clusters, whether the simple cluster detected with
\code{\link[=add_clusters]{add_clusters()}} or the
merged clusters detected with \code{\link[=merge_dynamic_clusters]{merge_dynamic_clusters()}}.}

\item{label_columns}{The column you want to be used to name the clusters. If the nodes are article,
you can choose, for instance, the columns with the author of the article and the
date of publication.}

\item{label_name}{The name of the column with cluster names, that will be created by the function.
"cluster_label" by default.}

\item{tidygraph_function}{For the \code{tidygraph_functions} method (see the details), the centrality measure to be
chosen among the measures implemented in in \code{tidygraph}  (see \code{tidygraph::centrality()}).}

\item{order_by}{For the \code{given_column} method, the column within the nodes list of your
tibble graph(s) you want to be used to classify nodes and choose names. This must be
a numeric column. For instance, you can use the \code{node_size} column of your network
if you have set \code{compute_size} to \code{TRUE} in \code{build_network()} or
\code{build_dynamic_networks()}.}

\item{text_columns}{For the \code{tf-idf} method, the columns with the text you want to analyze.
If you give multiple columns, they will be united to extract the terms. This is
a parameter of \code{extract_tfidf()}.}

\item{nb_terms_label}{For the \code{tf-idf} method, the number of terms you want to be used to serve a the
name of a cluster. Terms will be separated by a comma. This is a parameter of
\code{extract_tfidf()}.}

\item{...}{Additional arguments from \code{extract_tfidf}, outside of those referred above as well as
of \code{grouping_across_list} which is not relevant here.}
}
\value{
The same tibble graph or list of tibble graphs with a new column with the
names of the clusters, for both nodes and edges. If you choose the \code{tidygraph_functions} method,
the function also returns for nodes a column with the centrality measure computed.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

The function gives a name to networks clusters.
It also gives the edges the name of their cluster. The clusters are named according
to the column chosen by the user (for instance, in the case nodes are articles, the
name may be the author and date of an article).
}
\details{
The node to be used for naming the community is chosen depending on 3 methods:
\itemize{
\item the \code{tidygraph_functions} method: the name of a cluster comes from the node,
within the cluster, which has the highest centrality measure. The user can choose the
different centrality measure implemented  in \code{tidygraph} (see \code{tidygraph::centrality()} for details).
\item the \code{given_column} method: the user gives a column of the tibble graph(s), with numeric
values, that will be used to classify the nodes and choose the name of each cluster. The
\code{label_columns} of the node with the highest numerical value in the cluster will be used
to name the cluster.
\item the \code{tf-idf} method: clusters are name according to the terms with the highest
tf-idf value for each cluster. The user furnishes one or several columns with text,
and the function extracts the term and calculate the tf-idf value of each term, depending
on all the clusters. This method uses \code{extract_tfidf()}.
}

Please note that, when \code{name_merged_clusters} is set to \code{FALSE}, the TF-IDF is computed
tibble graph by tibble graph. It means that it is more likely that clusters in different
tibble graphs will share the same name.
}
\examples{
library(networkflow)

nodes <- Nodes_stagflation |>
dplyr::rename(ID_Art = ItemID_Ref) |>
dplyr::filter(Type == "Stagflation")

references <- Ref_stagflation |>
dplyr::rename(ID_Art = Citing_ItemID_Ref)

temporal_networks <- build_dynamic_networks(nodes = nodes,
directed_edges = references,
source_id = "ID_Art",
target_id = "ItemID_Ref",
time_variable = "Year",
cooccurrence_method = "coupling_similarity",
time_window = 20,
edges_threshold = 1,
overlapping_window = TRUE,
filter_components = TRUE,
verbose = FALSE)

temporal_networks <- add_clusters(temporal_networks,
objective_function = "modularity",
clustering_method = "leiden",
verbose = FALSE)

# You can name the clusters in each tibble graphs:

temporal_networks_with_names <- name_clusters(graphs = temporal_networks,
method = "tidygraph_functions",
name_merged_clusters = FALSE,
cluster_id = "cluster_leiden",
label_columns = c("Author", "Year"),
tidygraph_function = tidygraph::centrality_pagerank())

temporal_networks_with_names[[1]]

# Or you can name the dynamic clusters:

temporal_networks <- merge_dynamic_clusters(temporal_networks,
cluster_id = "cluster_leiden",
node_id = "ID_Art",
threshold_similarity = 0.51,
similarity_type = "partial")

temporal_networks_with_names <- name_clusters(graphs = temporal_networks,
method = "tf-idf",
name_merged_clusters = TRUE,
cluster_id = "dynamic_cluster_leiden",
text_columns = "Title",
nb_terms_label = 5,
clean_word_method = "lemmatise")

temporal_networks_with_names[[1]]

}
