% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks_to_alluv.R
\name{networks_to_alluv}
\alias{networks_to_alluv}
\title{Create a data.frame suitable for alluvial graph projection}
\usage{
networks_to_alluv(
  list_graph = NA,
  intertemporal_cluster_column = "intertemporal_name",
  node_key = NA,
  summary_cl_stats = FALSE
)
}
\arguments{
\item{list_graph}{Your list with all networks}

\item{intertemporal_cluster_column}{The column with the identifier of the inter-temporal cluster. If you have used
\link[=add_clusters]{add_clusters()} and \link[=merge_dynamic_clusters]{merge_dynamic_clusters()},
it is of the form \verb{dynamic_cluster_\{clustering_method\}}.}

\item{node_key}{The column with the unique identifier of each node.}

\item{summary_cl_stats}{If set to TRUE, the data.frame will contain a list of variable that summarize cluster statistics of the alluvial. These variables
can be particularly useful to filter smaller communities when plotting according to different variables:
\itemize{
\item \code{share_cluster_alluv} is the percentage share of a given cluster across all time windows;
\item \code{share_cluster_window} is the percentage share of a given cluster in a given time window;
\item \code{share_clusert_max} is the highest value of \code{share_cluster_window} for a given cluster across all individual time windows;
\item \code{length_cluster} is the number of time windows a cluster exists.
}

#' @examples}
}
\description{
This function creates a data.frame that can be easily plotted with ggalluvial from a list of networks.
}
