% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks_to_alluv.R
\name{networks_to_alluv}
\alias{networks_to_alluv}
\title{Create a data.frame suitable for alluvial graph projection}
\usage{
networks_to_alluv(
  list_graph = NA,
  cluster_column = NA,
  node_key = NA,
  summary_cl_stats = FALSE
)
}
\arguments{
\item{list_graph}{Your list with all networks}

\item{cluster_column}{The column with the identifier of the cluster.}

\item{node_key}{The column with the unique identifier of each node.}

\item{summary_cl_stats}{If set to TRUE, the data.frame will contain a list of variable that summarize cluster statistics of the alluvial. These variables
can be particularly useful to filter smaller communities when plotting according to different variables:
\itemize{
\item share_cl_alluv is the share of a given cluster across all time windows;
\item share_cl_window is the share of a given cluster in a given time window;
\item share_cl_max is the highest value of share_cl_window for a given cluster across all individual time windows;
\item length_cl is number of time windows a cluster exists.
}

#' @examples}
}
\description{
This function creates a data.frame that can be easily plotted with ggalluvial from a list of networks.
}
