% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_alluvial.R
\name{plot_alluvial}
\alias{plot_alluvial}
\title{Plot the alluvial after having followed the workflow of networkflow.}
\usage{
plot_alluvial(
  alluv_dt = NA,
  window_column = "Window",
  y_alluv_column = "y_alluv",
  stratum_column = "intertemporal_name",
  alluvium_column = "ID_Art",
  color = NA,
  label_column = "intertemporal_name_label",
  order_column = "minimize_crossing_order"
)
}
\arguments{
\item{alluv_dt}{Data.frame of the alluvial created using the networkflow::networks_to_alluv function}

\item{window_column}{The column with your time windows.}

\item{y_alluv_column}{The name of the column with y values created with \link[=networks_to_alluv]{networks_to_alluv()}}

\item{stratum_column}{The column with the identifier or label of the inter-temporal cluster. By default, "intertemporal_name", as it is
the name of the column created with \link[=intertemporal_cluster_naming]{intertemporal_cluster_naming()}.}

\item{alluvium_column}{The column with the unique identifier of each node.}

\item{color}{The colors you want to use in your alluvial
It can be a vector of colors or a two columns data.frame with the first column as the distinct observations of the stratum_column and a second column with the vector of colors you want to use}

\item{label_column}{The name of the column with labels created with \link[=label_alluvial]{label_alluvial()}}

\item{order_column}{The name of the column with the order of the stratum in a way that minimize crossing created with \link[=minimize_crossing_alluvial]{minimize_crossing_alluvial()}}
}
\description{
An easy way to plot the alluvial by using the columns created by the different function of networkflow.
}
